Added search spaces, moved DGP sampler, fixed bug, other minor fixes
