+++ trieste/models/gpflux/__init__.py
@@ -21,3 +21,4 @@ from .architectures import build_vanilla_deep_gp
 from .config import GPfluxModelConfig
 from .interface import GPfluxPredictor
 from .models import DeepGaussianProcess
+from .sampler import DeepGaussianProcessSampler
